Require zone and group on resources.
Looked like complexities from cascading attributes was creating issues.
We'd find that the zone would be removed after refreshes sometimes,
even though it wasn't coming down from upstream.

If it turns out we're wrong about that we can revert this.

It felt dangerous though to have implicit relationships for defaults.